updating tests for experimental always use trust for official images Signed-off-by: David Lawrence <david.lawrence@docker.com> (github: endophage)
